# qditp/docs/09-OpenAPI规范/fep-alipay-server-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "fep-alipay-server OpenAPI",
    "description": "支付宝出行 fep-alipay-server 入口网关接口文档，供测试使用。",
    "version": "2.0.4"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/api/payment/requestRefund": {
      "post": {
        "tags": [
          "支付宝出行-支付管理"
        ],
        "summary": "退款申请",
        "description": "接收支付宝出行侧退款申请。",
        "operationId": "requestRefund",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlipayTripRequestRefundReqDTO"
              },
              "example": {
                "orderNo": "ORDER202601010001",
                "cardIssueCode": "0007",
                "cardNum": "9900000000000001",
                "channelAgreementNo": "ALIPAY_AGT_001",
                "refundAmount": "200",
                "refundOrderNo": "REFUND202601010001"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "退款申请响应",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlipayTripRequestRefundRespDTO"
                },
                "example": {
                  "retCode": "0000",
                  "retMsg": "成功"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AlipayTripRequestRefundReqDTO": {
        "type": "object",
        "required": [
          "orderNo",
          "cardIssueCode",
          "cardNum",
          "channelAgreementNo",
          "refundAmount",
          "refundOrderNo"
        ],
        "properties": {
          "orderNo": {
            "type": "string",
            "description": "原订单号"
          },
          "cardIssueCode": {
            "type": "string",
            "description": "卡机构编号，支付宝0007"
          },
          "cardNum": {
            "type": "string",
            "description": "逻辑卡号"
          },
          "channelAgreementNo": {
            "type": "string",
            "description": "渠道协议号"
          },
          "refundAmount": {
            "type": "string",
            "description": "退款金额，单位分"
          },
          "refundOrderNo": {
            "type": "string",
            "description": "退款订单号"
          }
        }
      },
      "AlipayTripRequestRefundRespDTO": {
        "type": "object",
        "properties": {
          "retCode": {
            "type": "string",
            "description": "返回码"
          },
          "retMsg": {
            "type": "string",
            "description": "返回消息"
          }
        }
      },
...
    }
  }
}